requirements.txt
playwright==1.43.0
httpx==0.27.0
pyyaml==6.0.1
